Only include CTest if testing or fuzzing (#23)
When CTest is included, it adds a bunch of test-related targets that appear as projects in the MSVC solution (when using MSVC), e.g. Continuous, Experimental, Nightly, NightlyMemoryCheck.

A codebase that's just trying to use ValveFileVDF as a header-only library through CMake (e.g. using FetchContent) doesn't need these targets and they just clutter up the solution's list of projects, so only include CTest if testing or fuzzing is enabled.